ET: structure refine
