+++ lib/dashboard/model/sleep.dart
@@ -1,0 +1,3 @@
+class Sleep {
+  
+}
