+++ lib/dashboard/model/user.dart
