+++ lib/shared/bar_graph/model/inidividual_bar.dart
