--- lib/viewmodel/user_auth/authentication_model.dart
@@ -1,15 +1,0 @@
-// ignore_for_file: avoid_print
-
-import 'package:flutter/cupertino.dart';
-
-class AuthenticationViewModel extends ChangeNotifier {
-
-  void login(){
-    // Do something with the value
-    print("test");
-  }
-
-  void register(){
-    // Do something with the value
-  }
-}
